Change `points.size()` to `size()` to make transition to `index_t` easier (#4190)
* sed -i 's/\([^a-zA-Z0-9_]\)points.size/\1size/g'
* Replace all casts for `width = EXPR size()`
* Format string changes
* Auto
* Range based for loop
* OpenMP doesn't like range for loop

+++ filters/src/project_inliers.cpp
@@ -117,7 +117,7 @@ pcl::ProjectInliers<pcl::PCLPointCloud2>::applyFilter (PCLPointCloud2 &output)
     {
       // Copy everything
       output.height       = 1;
-      output.width        = static_cast<std::uint32_t> (indices_->size ());
+      output.width        = indices_->size ();
       output.point_step   = input_->point_step;
       output.data.resize (output.width * output.point_step);
       output.is_bigendian = input_->is_bigendian;
