+++ keypoints/include/pcl/keypoints/impl/trajkovic_3d.hpp
@@ -287,7 +287,7 @@ TrajkovicKeypoint3D<PointInT, PointOutT, NormalT>::detectKeypoints (PointCloudOu
   }
 
   output.height = 1;
-  output.width = static_cast<std::uint32_t> (output.size());
+  output.width = output.size();
   // we don not change the denseness
   output.is_dense = true;
 }
